+++ pesacheck_meedan_bridge/.env.example
@@ -1,6 +1,6 @@
 PESACHECK_URL=https://pesacheck.org/feed
-PESACHECK_RSS_2_JSON_API_KEY=
+PESACHECK_RSS2JSON_API_KEY=
 PESACHECK_CHECK_URL=https://check-api.checkmedia.org/api/graphql
 PESACHECK_TOKEN=
-PESACHECK_WORKSPACE_SLUG=pesacheck-tipline-sandbox
+PESACHECK_CHECK_WORKSPACE_SLUG=pesacheck-tipline-sandbox
 PESACHECK_SENTRY_DSN=
